integrate addCsv within writeToCsv
