+++ src/app.ts
@@ -103,7 +103,6 @@ async function scrapeAndProcess() {
 
 /* TODO:
 - fix issue with npm start
-- find way to automatically import files to anki
 - parse content to remove whitespace and css code
 - truncate content and question to 4000 characters / 500 words due to chatgpt limits
 */
