+++ src/writeToCsv.ts
@@ -2,6 +2,8 @@ import { fileURLToPath } from "url";
 import { dirname } from "path";
 import fs from "fs";
 import { createObjectCsvWriter as createCsvWriter } from "csv-writer";
+import csv from "csv-parser";
+import axios from "axios";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
@@ -23,13 +25,13 @@ type writeToCsvProps = {
   }[];
 };
 
-function writeToCsv({
+async function writeToCsv({
   url,
   title,
   date,
   vocab,
   sentences,
-}: writeToCsvProps): void {
+}: writeToCsvProps): Promise<void> {
   // create folder if it doesn't exist
   if (!fs.existsSync(dataPath)) {
     fs.mkdirSync(dataPath);
@@ -37,6 +39,7 @@ function writeToCsv({
 
   // If you don't want to write a header line, don't give title to header elements and just give field IDs as a string.
   // removing titles for now
+  // Set up CSV writer for vocab
   let csvWriter = createCsvWriter({
     path: `${dataPath}/vnexpress_${date}_vocab.csv`,
     header: ["VN_word", "EN_word", "roots", "title", "date", "url"],
@@ -58,11 +61,17 @@ function writeToCsv({
     };
   });
 
-  csvWriter
-    .writeRecords(vocabData)
-    .then(() => console.log("The vocab CSV file was written successfully"))
-    .catch((err: Error) => console.log(err));
+  let fileName = `vnexpress_${date}_vocab.csv`;
+  try {
+    await csvWriter.writeRecords(vocabData);
+    console.log(`${fileName} was written successfully`);
+    await addCsvToAnki(fileName);
+    console.log(`${fileName} was added to Anki successfully`);
+  } catch (err) {
+    console.log(err);
+  }
 
+  // Set up CSV writer for sentences
   csvWriter = createCsvWriter({
     path: `${dataPath}/vnexpress_${date}_sentences.csv`,
     header: ["VN_sentence", "EN_sentence", "title", "date", "url"],
@@ -77,10 +86,89 @@ function writeToCsv({
     };
   });
 
-  csvWriter
-    .writeRecords(sentencesData)
-    .then(() => console.log("The sentences CSV file was written successfully"))
-    .catch((err: Error) => console.log(err));
+  fileName = `vnexpress_${date}_sentences.csv`;
+  try {
+    await csvWriter.writeRecords(sentencesData);
+    console.log(`${fileName} was written successfully`);
+    await addCsvToAnki(fileName);
+    console.log(`${fileName} was added to Anki successfully`);
+  } catch (err) {
+    console.log(err);
+  }
+}
+
+async function addCsvToAnki(fileName: string) {
+  const data: any[] = [];
+  const __filename = fileURLToPath(import.meta.url);
+  const __dirname = dirname(__filename);
+  const filePath = `${__dirname}/data/${fileName}`;
+
+  // check if fileName includes "vocab" or "sentences"
+  let headers: string[];
+  let deckName: string;
+  let modelName: string;
+  if (fileName.includes("vocab")) {
+    headers = ["VN_word", "EN_word", "roots", "title", "date", "url"];
+    deckName = "Tiếng Việt: Chung::Tin tức::Vocab";
+    modelName = "News-Vocab";
+  } else if (fileName.includes("sentences")) {
+    headers = ["VN_sentence", "EN_sentence", "title", "date", "url"];
+    deckName = "Tiếng Việt: Chung::Tin tức::Sentences";
+    modelName = "News-Sentences";
+  } else {
+    throw new Error("Invalid file name");
+  }
+
+  // Read the CSV file
+  fs.createReadStream(filePath)
+    .pipe(
+      csv({
+        separator: ";",
+        headers: headers,
+        skipLines: 0,
+      })
+    )
+    .on("data", (row: string[]) => {
+      // Push each row of data to an array
+      data.push(row);
+    })
+    .on("end", async () => {
+      // Iterate over the data array, destructuring each object
+      for (const row of data) {
+        // Create an empty object to store the field values
+        const fields: { [key: string]: string } = {};
+
+        // Map the header names to the corresponding field values
+        for (let i = 0; i < headers.length; i++) {
+          const header = headers[i];
+          fields[header] = row[header];
+        }
+
+        // Prepare the payload to add the note to Anki
+        const payload = {
+          action: "addNote",
+          version: 6,
+          params: {
+            note: {
+              deckName,
+              modelName,
+              fields,
+              options: {
+                allowDuplicate: true,
+              },
+              tags: [],
+            },
+          },
+        };
+
+        try {
+          // Send the HTTP POST request to AnkiConnect API
+          await axios.post("http://127.0.0.1:8765", payload);
+        } catch (error) {
+          console.error("An error occurred while adding note:", error);
+        }
+      }
+    });
 }
 
 export default writeToCsv;
